Server_API/requirements.txt
Flask
PyJWT
pymongo
